fix code review comment
